Split DTO and Entity packages again
