+++ src/main/java/uk/gov/ons/census/caseapisvc/model/dto/RefusalTypeDTO.java
@@ -1,0 +1,6 @@
+package uk.gov.ons.census.caseapisvc.model.dto;
+
+public enum RefusalTypeDTO {
+  HARD_REFUSAL,
+  EXTRAORDINARY_REFUSAL
+}
